requirements.txt
Django~=2.2.24
djangorestframework~=3.12.4
django-extensions
python-dotenv~=0.19.1
whitenoise~=5.3.0
gunicorn
psycopg2
dj_database_url
